[refactor] moved answer service to own package
