+++ src/main/java/com/screamer/resume/service/answer/AnswerDbService.java
@@ -1,4 +1,4 @@
-package com.screamer.resume.service.message;
+package com.screamer.resume.service.answer;
 
 import com.screamer.resume.entity.Message;
 import com.screamer.resume.exceptions.MessageNotFoundException;
