+++ src/test/java/com/screamer/resume/service/message/AnswerDbServiceTest.java
@@ -5,6 +5,7 @@ import com.screamer.resume.entity.Message;
 import com.screamer.resume.exceptions.MessageNotFoundException;
 import com.screamer.resume.exceptions.MessageUnansweredException;
 import com.screamer.resume.repository.MessageRepository;
+import com.screamer.resume.service.answer.AnswerDbServiceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
